+++ EP Mobile/EP Mobile/EPSSimpleAlgorithmViewController.m
@@ -175,7 +175,7 @@
 }
 
 - (void)alertView:(UIAlertView *)actionSheet clickedButtonAtIndex:(NSInteger)buttonIndex {
-    NSLog(@"Button index = %d", buttonIndex);
+    NSLog(@"Button index = %ld", (long)buttonIndex);
     // show map button
     if (buttonIndex == 1) {
         [self performSegueWithIdentifier:@"MapSegue" sender:nil];
